fix the scripts that still would fail on x86_64/i386
in particular, moonlight flatpack was added for x86_64
minecraft java multimc uses the normal repo for x86_64
(I think the binaries should also work fine on generic aarch64 so I've left that open as available to install)

+++ scripts/moonlight.sh
@@ -3,20 +3,33 @@
 cd
 clear -x
 echo "Moonlight script started!"
-sleep 1
-# remove old moonlight name and repo
-sudo rm -rf  /etc/apt/sources.list.d/moonlight-l4t.list
-sudo apt remove moonlight -y
+# get architecture using function
+get_system
+if [[ $jetson_model ]]; then
+    # remove old moonlight name and repo
+    sudo rm -rf  /etc/apt/sources.list.d/moonlight-l4t.list
+    sudo apt remove moonlight -y
 
-curl -1sLf 'https://dl.cloudsmith.io/public/moonlight-game-streaming/moonlight-l4t/setup.deb.sh' | sudo -E bash
-# replace apt source to bionic as its the only one hosted (allows moonlight to be installed on ubuntu focal, hirsute, etc on switch)
-sudo sed -i 's/ubuntu.*/ubuntu bionic main/' '/etc/apt/sources.list.d/moonlight-game-streaming-moonlight-l4t.list'
-sudo apt install moonlight-qt -y
+    curl -1sLf 'https://dl.cloudsmith.io/public/moonlight-game-streaming/moonlight-l4t/setup.deb.sh' | sudo -E bash
+    # replace apt source to bionic as its the only one hosted (allows moonlight to be installed on ubuntu focal, hirsute, etc on switch)
+    sudo sed -i 's/ubuntu.*/ubuntu bionic main/' '/etc/apt/sources.list.d/moonlight-game-streaming-moonlight-l4t.list'
+    sudo apt install moonlight-qt -y
+else
+    if [[ $architecture == "x86_64" ]]; then
+        sudo add-apt-repository ppa:alexlarsson/flatpak -y
+        sudo apt update
+        sudo apt install flatpak -y
+        flatpak remote-add --if-not-exists flathub https://flathub.org/repo/flathub.flatpakrepo
+        flatpak install flathub com.moonlight_stream.Moonlight
+    else
+        echo "Error: your cpu architecture ($architecture) is not supporeted by Moonlight (or we don't know how to install) and will fail to run"; echo ""; echo "Exiting the script"; sleep 3; exit $?
+    fi
+fi
 
 echo "Done!"
 echo "Ctrl + click this link before this message disappears in 20 seconds"
 echo "For a guide on how to set Moonlight up on your PC and connect"
-echo "to it from your Switch!"
+echo "to it from your $jetson_model $architecture device!"
 echo
 echo "https://github.com/moonlight-stream/moonlight-docs/wiki/Setup-Guide"
 echo
